+++ geolocator_apple/lib/src/geolocator_apple.dart
@@ -234,6 +234,8 @@ class GeolocatorApple extends GeolocatorPlatform {
         return const LocationServiceDisabledException();
       case 'LOCATION_SUBSCRIPTION_ACTIVE':
         return const AlreadySubscribedException();
+      case 'LOCATION_SIGNAL_LOST':
+        return const LocationSignalLostException();
       case 'PERMISSION_DEFINITIONS_NOT_FOUND':
         return PermissionDefinitionsNotFoundException(exception.message);
       case 'PERMISSION_DENIED':